Middleware admin, profesor y alumno
Para evitar ingreso no autorizado a secciones de la plataforma con roles equivocados.

+++ web/routes/web.php
@@ -11,56 +11,52 @@
 |
 */
 
-Route::get('/', function () {
-	return view('index');
-})->middleware('auth');
 Route::resource('/', 'IndexController')->middleware('auth');
-
 Route::resource('/empresas', 'EmpresaController')->middleware('auth');
 
 
 //Rutas de administración
 //Index
-Route::get('/admin', 'AdminController@index')->name('admin.index')->middleware('auth');
+Route::get('/admin', 'AdminController@index')->name('admin.index')->middleware('auth', 'admin');
 //Estadísticas de pasantías (TODO)
-Route::resource('/admin/estadisticas', 'GraficasController')->middleware('auth');
+Route::resource('/admin/estadisticas', 'GraficasController')->middleware('auth', 'admin');
 //Importar listado de alumnos autorizados para usar la plataforma
-Route::resource('/admin/importarlista', 'ListadoController')->middleware('auth');
+Route::resource('/admin/importarlista', 'ListadoController')->middleware('auth', 'admin');
 //Asignar alumnos a los profesores correspondientes
-Route::get('/admin/asignarProyectos', 'AdminController@asignarProyectosView')->name('admin.asignarProyectos')->middleware('auth');
-Route::get('/admin/asignarProyectos/{id}', 'AdminController@asignarProyectosManual')->middleware('auth');
-Route::get('/admin/asignarProyectos/{idProf}/{idProy}/{action}', 'AdminController@asignarProyectoQuick')->middleware('auth');
+Route::get('/admin/asignarProyectos', 'AdminController@asignarProyectosView')->name('admin.asignarProyectos')->middleware('auth', 'admin');
+Route::get('/admin/asignarProyectos/{id}', 'AdminController@asignarProyectosManual')->middleware('auth', 'admin');
+Route::get('/admin/asignarProyectos/{idProf}/{idProy}/{action}', 'AdminController@asignarProyectoQuick')->middleware('auth', 'admin');
 
 // Rutas de Listado Inscripcion
 // Ruta de Destroy Paso 2
-Route::post('/admin/listadoInscripcion/{id}/edit/paso2D/', 'ListadoInscripcionController@destroyPaso2')->name('listadoInscripcion.destroyPaso2')->middleware('auth');
+Route::post('/admin/listadoInscripcion/{id}/edit/paso2D/', 'ListadoInscripcionController@destroyPaso2')->name('listadoInscripcion.destroyPaso2')->middleware('auth', 'admin');
 // Ruta de Destroy Paso 3
-Route::post('/admin/listadoInscripcion/{id}/edit/paso3D/', 'ListadoInscripcionController@destroyPaso3')->name('listadoInscripcion.destroyPaso3')->middleware('auth');
+Route::post('/admin/listadoInscripcion/{id}/edit/paso3D/', 'ListadoInscripcionController@destroyPaso3')->name('listadoInscripcion.destroyPaso3')->middleware('auth', 'admin');
 // Ruta de Update Paso 2
-Route::post('/admin/listadoInscripcion/{id}/edit/paso2E/', 'ListadoInscripcionController@updatePaso2')->name('listadoInscripcion.updatePaso2')->middleware('auth');
+Route::post('/admin/listadoInscripcion/{id}/edit/paso2E/', 'ListadoInscripcionController@updatePaso2')->name('listadoInscripcion.updatePaso2')->middleware('auth', 'admin');
 //Ruta de Update Paso 3
-Route::post('/admin/listadoInscripcion/{id}/edit/paso3E/', 'ListadoInscripcionController@updatePaso3')->name('listadoInscripcion.updatePaso3')->middleware('auth');
+Route::post('/admin/listadoInscripcion/{id}/edit/paso3E/', 'ListadoInscripcionController@updatePaso3')->name('listadoInscripcion.updatePaso3')->middleware('auth', 'admin');
 // Ruta de editado de empresa unicamente
-Route::get('/admin/listadoInscripcion/{id}', 'EmpresaController@edit')->middleware('auth');
+Route::get('/admin/listadoInscripcion/{id}', 'EmpresaController@edit')->middleware('auth', 'admin');
 // Ruta de CRUD Listado Inscripcion
-Route::resource('/admin/listadoInscripcion', 'ListadoInscripcionController')->middleware('auth');
+Route::resource('/admin/listadoInscripcion', 'ListadoInscripcionController')->middleware('auth', 'admin');
 // Ruta de exportacion de excel
-Route::get('/admin/tablaInscripciones', 'ListadoInscripcionController@export')->name( 'tablaInscripciones.export')->middleware('auth');
+Route::get('/admin/tablaInscripciones', 'ListadoInscripcionController@export')->name( 'tablaInscripciones.export')->middleware('auth', 'admin');
 // Ruta de validar al pariente
-Route::get('/admin/listadoInscripcion/{id}/statusPaso2/{statusPaso2}', 'ListadoInscripcionController@validarPariente')->name('listadoInscripcion.validarPariente')->middleware('auth');
+Route::get('/admin/listadoInscripcion/{id}/statusPaso2/{statusPaso2}', 'ListadoInscripcionController@validarPariente')->name('listadoInscripcion.validarPariente')->middleware('auth', 'admin');
 // Ruta de validar proyecto
-Route::get('/admin/listadoInscripcion/{id}/accion/{accion}', 'ListadoInscripcionController@validarProyecto')->name('listadoInscripcion.validarProyecto')->middleware('auth');
+Route::get('/admin/listadoInscripcion/{id}/accion/{accion}', 'ListadoInscripcionController@validarProyecto')->name('listadoInscripcion.validarProyecto')->middleware('auth', 'admin');
 // Ruta de validar todo
-Route::get('/admin/listadoInscripcion/{nombresUsuario}/idPasantia/{idPasantia}', 'ListadoInscripcionController@validarTodo')->name('listadoInscripcion.validarTodo')->middleware('auth');
+Route::get('/admin/listadoInscripcion/{nombresUsuario}/idPasantia/{idPasantia}', 'ListadoInscripcionController@validarTodo')->name('listadoInscripcion.validarTodo')->middleware('auth', 'admin');
 
 
 
 
 Route::resource('/perfil', 'PerfilController')->middleware('auth');
 
-Route::get('/profesor', 'ProfesorController@index')->name('profesor.index')->middleware('auth');
-Route::get('/profesor/proyecto/{id}', 'ProfesorController@verProyecto')->name('profesor.verProyecto')->middleware('auth');
-Route::post('/profesor/proyecto/{id}/feedback', 'ProfesorController@feedbackProyecto')->middleware('auth');
+Route::get('/profesor', 'ProfesorController@index')->name('profesor.index')->middleware('auth', 'noAlumno');
+Route::get('/profesor/proyecto/{id}', 'ProfesorController@verProyecto')->name('profesor.verProyecto')->middleware('auth', 'noAlumno');
+Route::post('/profesor/proyecto/{id}/feedback', 'ProfesorController@feedbackProyecto')->middleware('auth', 'noAlumno');
 
 
 
@@ -90,10 +86,10 @@ Route::get('/inscripcion/certificado', 'PasantiaController@descargarCert')->name
 
 Route::get('/confirmarTutor/{id}', 'PasantiaController@confirmarTutor')->name('confTutor');
 
-Route::delete('/inscripcion/destroy/{id}','PasantiaController@destroy')->name('inscripcion.destroy')->middleware('auth');
+Route::delete('/inscripcion/destroy/{id}','PasantiaController@destroy')->name('inscripcion.destroy')->middleware('auth', 'admin');
 
 Route::post('evaluacion/{id}','EvalTutorController@save')->name('evalTutor.save');
 Route::get('evaluacion/{id}', 'EvalTutorController@show')->name('evalTutor.show');
 Route::get('evaluacion/enviar/{idAlumno}', 'EvalTutorController@enviar');
-Route::get('evaluacion/listado/{idProyecto}', 'EvalTutorController@listado')->name('EvalTutor.listado')->middleware('auth');
-Route::get('evaluacion/ver/{idEvaluacion}', 'EvalTutorController@ver')->middleware('auth');
+Route::get('evaluacion/listado/{idProyecto}', 'EvalTutorController@listado')->name('EvalTutor.listado')->middleware('auth', 'noAlumno');
+Route::get('evaluacion/ver/{idEvaluacion}', 'EvalTutorController@ver')->middleware('auth', 'noAlumno');
